Put prepocess.py code in dataset.py to adhere to assignment specs
